+++ esbuild.config.mjs
@@ -40,6 +40,9 @@ const context = await esbuild.context({
     treeShaking: true,
     outfile: 'main.js',
     minify: prod,
+    define: {
+        __DEV__: prod ? 'false' : 'true',
+    },
 });
 
 if (prod) {